requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asyncio-dgram==2.1.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cloudscraper==1.2.60
colorama==0.4.5
dnspython==2.2.1
dropbox==11.33.0
frozenlist==1.3.1
idna==3.3
Js2Py==0.71
lxml==4.9.1
mcstatus==9.2.0
multidict==6.0.2
pip==22.0.4
ply==3.11
py-cord==2.0.0
pyjsparser==2.7.1
pyodbc==4.0.34
pyparsing==3.0.9
python-aternos==1.1.1
python-dotenv==0.20.0
pytz-deprecation-shim==0.1.0.post0
regex==2022.7.25
requests==2.28.1
requests-toolbelt==0.9.1
setuptools==58.1.0
six==1.16.0
stone==3.3.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.11
websockets==10.3
yarl==1.8.1
